Moved states list to constants
